IA/ai-docs-translator/requirements.txt
python-dotenv
pdfplumber
requests
azure-core
azure-ai-textanalytics
gradio